backend/requirements.txt
fastapi==0.116.2
uvicorn==0.35.0
supabase==2.19.0
textblob==0.19.0
python-dotenv==1.1.1
requests==2.32.5
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.3
nltk==3.9.1
pydantic==2.11.9
python-multipart==0.0.12
